Fix compiler warnings in DirectX11 sources
Clears every warning MSVC reports for our own sources (DirectXTK and
SDK headers excluded) so new ones stand out:

- DEFINE_OPERATOR bodies return int/double from a float evaluate();
  the conversion is the operator contract, made explicit with one
  cast in the macro (C4244 x12).
- %d used with size_t in SetDataCacheRegion logs (C4477).
- GetIniInt called with -1.0f defaults for int settings.
- Explicit casts at the remaining narrowing/sign-mismatch sites
  (size_t/UINT64/uint64 -> int/UINT, std::distance -> int, sentinel
  -1 comparisons) and TargetInfo::max_slot_count made unsigned.
- Three unreferenced locals removed, _wfopen -> _wfopen_s.
- GetElementIndex: default: assert(false) fell off the end in release
  builds; return SIZE_MAX (C4715).
- KeyOverride gets the same C4250 suppress KeyOverrideCycle has.

+++ DirectX11/FrameAnalysis.cpp
@@ -863,7 +863,7 @@ void FrameAnalysisContext::dedupe_buf_filename_vb_txt(const wchar_t *bin_filenam
 	StringCchPrintfExW(pos, rem, &pos, &rem, NULL, L"-vb%i", idx);
 
 	if (layout) {
-		uint32_t layout_hash = layout->GetLayoutHash();
+		uint32_t layout_hash = (uint32_t)layout->GetLayoutHash();
 		StringCchPrintfExW(pos, rem, &pos, &rem, NULL, L"-layout=%08x", layout_hash);
 	}
